backend/requirements.txt
requests
pillow
opencv-python
fastapi
uvicorn[standard]
websockets
serpapi
python-dotenv
python-multipart
yt-dlp
google-generativeai
google-auth
google-auth-oauthlib
google-auth-httplib2
scikit-learn
numpy
pymongo
upstash-redis
google-search-results
cloudinary
torch 
transformers 
pytorchvideo